+++ src/main/java/com/app/entities/enums/GenderEnum.java
@@ -1,4 +1,4 @@
-package com.app.entities;
+package com.app.entities.enums;
 
 public enum GenderEnum {
