+++ src/main/java/com/app/entities/enums/TitleEnum.java
@@ -1,4 +1,4 @@
-package com.app.entities;
+package com.app.entities.enums;
 
 public enum TitleEnum {
 	Mr,Mrs,Dr,Miss,Prof
