+++ src/main/resources/application.properties
@@ -5,8 +5,8 @@
 server.servlet.session.persistent=false
 #DB properties
 spring.datasource.url=jdbc:mysql://localhost:3306/crime_portal?createDatabaseIfNotExist=true&useSSL=false&allowPublicKeyRetrieval=true
-spring.datasource.username=D6_80514_Praneet
-spring.datasource.password=sunbeam
+spring.datasource.username=project
+spring.datasource.password=project
 # JPA properties
 spring.jpa.show-sql = true
 spring.jpa.hibernate.ddl-auto = update
